+++ Automate/Framework/Machines/Objects/LoomMachine.cs
@@ -5,6 +5,7 @@ using SObject = StardewValley.Object;
 namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
 {
     /// <summary>A loom that accepts input and provides output.</summary>
+    /// <remarks>Derived from <see cref="SObject.performObjectDropInAction"/>.</remarks>
     internal class LoomMachine : GenericObjectMachine<SObject>
     {
         /*********
@@ -50,7 +51,46 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
         /// <returns>Returns whether the machine started processing an item.</returns>
         public override bool SetInput(IStorage input)
         {
-            return this.GenericPullRecipe(input, this.Recipes);
+            if (input.TryGetIngredient(this.Recipes, out IConsumable consumable, out IRecipe recipe))
+            {
+                // get output
+                SObject output = recipe.Output(consumable.Take());
+                if (consumable.Sample is SObject sampleInput)
+                {
+                    if (Game1.random.NextDouble() <= this.GetProbabilityOfDoubleOutput(sampleInput.Quality))
+                        output.Stack = 2;
+                }
+
+                this.Machine.heldObject.Value = output;
+                this.Machine.MinutesUntilReady = recipe.Minutes;
+                return true;
+            }
+
+            return false;
+        }
+
+
+        /*********
+        ** Private methods
+        *********/
+        /// <summary>Get the probability that the cheese press outputs two items for a given input quality.</summary>
+        /// <param name="quality">The input quality.</param>
+        private float GetProbabilityOfDoubleOutput(int quality)
+        {
+            switch (quality)
+            {
+                case SObject.lowQuality:
+                    return 0;
+
+                case SObject.highQuality:
+                    return 0.25f;
+
+                case SObject.bestQuality:
+                    return 0.5f;
+
+                default:
+                    return 0.1f;
+            }
         }
     }
 }